rollback datafiles loader
